moved to .net9 since it has built in uuid V7; added some domain models, added database inital migrations,

+++ backend/src/Infrastructure/Services.Implementation/OSM/Models/LatLngJsonModel.cs
@@ -5,8 +5,8 @@ namespace Services.Implementation.OSM.Models;
 internal sealed record LatLngJsonModel
 {
     [JsonPropertyName("lat")]
-    public decimal Lat { get; init; }
+    public double Lat { get; init; }
 
     [JsonPropertyName("lon")]
-    public decimal Lng { get; init; }
+    public double Lng { get; init; }
 }
